backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
ultralytics>=8.3.0
python-multipart==0.0.6
pillow>=11.0.0
python-dotenv==1.0.0
pydantic==2.5.3
numpy>=1.26.0,<2.0
cloudinary>=1.36.0
